refactor(sampling): move plan inspection into its own translation unit
Give SamplingPlan the same file layout as ExecutablePlan: the IR and its
validation stay in plan.cc while the inspection text rendering moves to
plan_inspection.cc. A pure move; the characterization digest is unchanged.
